OpNav subprocess move to FM WIP
