+++ flight_modes/opnav_flightmode.py
@@ -2,11 +2,15 @@ from datetime import datetime
 from time import sleep
 import sqlite3
 from sqlalchemy.exc import SQLAlchemyError
+from multiprocessing import Process, Queue
+import subprocess
 
 from utils.constants import FMEnum
 from utils.db import OpNavCoordinatesModel
+from utils.log import get_log
 from .flight_mode import PauseBackgroundMode
 
+logger = get_log()
 
 class OpNavMode(PauseBackgroundMode):
 
@@ -62,6 +66,12 @@ class OpNavMode(PauseBackgroundMode):
                 raise Exception("Error while reading most recent OpNav result")
 
     def run_mode(self):
+        if not self.opnav_process.is_alive():
+            logger.info("[OPNAV]: Able to run next opnav")
+            self.parent.last_opnav_run = datetime.now()
+            logger.info("[OPNAV]: Starting opnav subprocess")
+            self.opnav_process = Process(target=self.opnav_subprocess, args=())
+            self.opnav_process.start()
         self.run_opnav()
         self.task_completed()
 
@@ -74,3 +84,11 @@ class OpNavMode(PauseBackgroundMode):
         # return to normal mode if task completed
         if self.task_completed:
             return FMEnum.Normal.value
+
+
+    def opnav_subprocess(self):
+        # TODO change from pytest to actual opnav
+        # note: os.system should be replaced by subprocess.run("<insert shell command here>", shell=True)
+        # for an example see utils/boot_cause.py
+        # os.system("pytest OpticalNavigation/tests/test_pipeline.py::test_start")
+        subprocess.run('pytest OpticalNavigation/tests/test_pipeline.py::test_start', shell=True)
